some core changes

+++ xlxd_dash/xlxd.py
@@ -7,26 +7,28 @@ def xlxd_data():
     res = {}
     res['running'] = False
 
-    pid = 0
     pid_file = config["xlxd"]["pidfile"]
-    res["uptime"] = int(time.time() - os.path.getctime(pid_file))
+    res["uptime"] = 0
+    if os.path.exists(pid_file):
+        res["uptime"] = int(time.time() - os.path.getctime(pid_file))
 
-    with open(pid_file) as f:
-        pid = f.read()
+        with open(pid_file) as f:
+            pid = f.read()
 
-    try:
-        os.kill(int(pid), 0)
-        res['running'] = True
-    except:
-        pass
+        try:
+            os.kill(int(pid), 0)
+            res['running'] = True
+        except Exception as e:
+            print(f"XLXD process check failed: {e}")
 
     xml_file = config["xlxd"]["xml"]
-    with open(xml_file) as f:
-        contenido = f.read()
-
     res["version"] = 'N/A'
-    match = re.search(r"<Version>(.*?)</Version>", contenido)
-    if match:
-        res["version"] = match.group(1)
+    if os.path.exists(xml_file):
+        with open(xml_file) as f:
+            contenido = f.read()
+
+        match = re.search(r"<Version>(.*?)</Version>", contenido)
+        if match:
+            res["version"] = match.group(1)
 
     return res
